split widgets, add exports
