+++ src/aws-provider/aws-credentials/index.ts
@@ -1,0 +1,4 @@
+export * from './aws-assumed-role.js';
+export * from './aws-credentials-type.js';
+export * from './aws-keys.js';
+export * from './local-aws-profile.js';
