+++ src/node.ts
@@ -1,7 +1,6 @@
-export * as Controllers from './controllers/index.js';
 export * as Models from './models/index.js';
-export * as Views from './views/index.js';
 export * as OpsTypes from './ops-types.js';
+export * from './controllers/index.js';
 export * from './utils/index.js';
 export * from './types/types.js';
 export * from './aws-provider/aws-credentials-provider.js';
